Nit in RPC.cpp

+++ RPC.h
@@ -346,7 +346,7 @@ namespace RPC {
         QByteArray wrapForSend(const QByteArray &) override;
     private:
         bool memoryWasteTimerActive = false;  ///< inticates the DoS protection timer is active, used by memoryWasteDoSProtection
-        qint64 memoryWasteThreshold = -1; ///< gets lazy-initialized in memoryWasteDoSProtection below
+        int memoryWasteThreshold = -1; ///< gets lazy-initialized in memoryWasteDoSProtection below
         void memoryWasteDoSProtection(); ///< must be called from on_readyRead only.
     };
